fix: fix paths and warnings
