+++ src/components/ui/button/button.tsx
@@ -1,6 +1,6 @@
 import { forwardRef } from "react";
 import { Slot } from "@radix-ui/react-slot";
-import { cva, type VariantProps } from "class-variance-authority";
+import { type VariantProps } from "class-variance-authority";
 import { cn } from "@utils/utils";
 import { renderIcon } from "@utils/renders";
 
@@ -24,6 +24,7 @@ import message from "@assets/menu/message.svg";
 import news from "@assets/menu/news.svg";
 import photos from "@assets/menu/photos.svg";
 import user from "@assets/menu/user.svg";
+import { buttonVariants } from "../variants/variants";
 
 const icons = {
   edit,
@@ -44,33 +45,6 @@ const icons = {
   message,
   photos,
 };
-
-export const buttonVariants = cva(
-  "w-full inline-flex items-center justify-center gap-[5px] rounded-md text-sm font-medium focus-visible:outline-none disabled:pointer-events-none disabled:opacity-50 transition-all active:scale-[0.95]",
-  {
-    variants: {
-      variant: {
-        default: "bg-primary text-primary-foreground hover:bg-primary-hover",
-        secondary:
-          "bg-secondary text-secondary-foreground border border-secondary-border hover:border-secondary-borderHover hover:bg-secondary-hover",
-        outline:
-          "text-outline-border border border-outline-border bg-background hover:border-outline-hover hover:text-outline-hover",
-        link: "text-link-foreground underline-offset-4 hover:text-link-hover",
-        text: "text-text-foreground hover:bg-text",
-      },
-      size: {
-        default: "h-10 px-[0.95rem] py-[0.625rem]",
-        sm: "h-9 rounded-md px-3",
-        lg: "h-11 rounded-md px-8",
-        icon: "h-10 w-10",
-      },
-    },
-    defaultVariants: {
-      variant: "default",
-      size: "default",
-    },
-  },
-);
 
 export type Icon = keyof typeof icons;
