+++ src/pages/not-found-page/not-found-page.tsx
@@ -1,7 +1,7 @@
+import { FC } from "react";
 import notFoundDark from "@assets/404 dark.svg";
 import notFoundLight from "@assets/404 light.svg";
-import { useTheme } from "@src/components/theme-provider/theme-provider";
-import { FC } from "react";
+import { useTheme } from "@src/hooks/useTheme";
 
 const NotFoundPage: FC<{ type: "inside" | "outside" }> = ({ type }) => {
   const { theme } = useTheme();
