+++ src/pages/registration/registration.tsx
@@ -6,9 +6,9 @@ import { registrationThunk } from "@src/store/reducers/profileInfo/async-thunks"
 import { useAppDispatch } from "@src/hooks/store-hooks";
 import Logo from "@src/components/ui/logo/logo";
 import SettingButton from "@src/components/ui/button/setting-button/setting-button";
-import { useTheme } from "@src/components/theme-provider/theme-provider";
 import Input from "@src/components/ui/form/input/input";
 import Button from "@src/components/ui/button/button";
+import { useTheme } from "@src/hooks/useTheme";
 
 const RegistrationPage: FC = () => {
   const dispatch = useAppDispatch();
